# perf: batch wallets on insert