## Proposed changes
<!--
Describe the big picture of your changes here to communicate to the maintainers why we should accept this pull request. If it fixes a bug or resolves a feature request, be sure to link to that issue.
-->
By properly batching the wallets when doing a force insert we can cut down the time spent considerably. Increasing the chunk size has no visible effect, so I left it unchanged.

Before (devnet):
```
saveWallets: 2456.495ms
saveWallets: 2487.363ms
saveWallets: 2148.759ms
saveWallets: 2336.846ms
saveWallets: 1850.622ms
saveWallets: 1862.646ms
saveWallets: 1843.057ms
saveWallets: 1936.318ms
saveWallets: 1806.108ms
saveWallets: 1906.639ms
```

After (devnet):
```
saveWallets: 1190.114ms
saveWallets: 1162.871ms
saveWallets: 1218.505ms
saveWallets: 1288.276ms
saveWallets: 1119.453ms
saveWallets: 1365.919ms
saveWallets: 1572.861ms
saveWallets: 1252.263ms
saveWallets: 1317.798ms
saveWallets: 1177.376ms
```


## Types of changes
<!--
What types of changes does your code introduce?
_Put an `x` in the boxes that apply and remove the rest of them: keep this PR as concise, but descriptive, as possible.._
-->

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [X] Refactoring (improve a current implementation without adding a new feature or fixing a bug)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Build (changes that affect the build system)
- [ ] Docs (documentation only changes)
- [ ] Test (adding missing tests or fixing existing tests)
- [ ] Other

## Checklist
<!--
_Put an `x` in the boxes that apply and remove this text and the rest of them. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code._
-->

- [ ] I have read the [CONTRIBUTING](https://docs.ark.io/developers/guidelines/contributing.html) documentation
- [X] Lint and unit tests pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
<!--

## Further comments

If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc...
-->